Move route logic to controller
